+++ framework/py/flwr/supercore/cli/flwr_model.py
@@ -23,11 +23,11 @@ from flwr.common.args import add_args_flwr_app_common, try_obtain_flwr_app_token
 from flwr.common.constant import SERVERAPPIO_API_DEFAULT_CLIENT_ADDRESS
 from flwr.common.exit import ExitCode, flwr_exit
 from flwr.common.logger import log, mirror_output_to_queue, restore_output
-from flwr.supercore.executors.run_model import run_model
+from flwr.supercore.executors.model.run import run_model
 
 
 def flwr_model() -> None:
-    """Run process-isolated Flower ModelApp."""
+    """Run process-isolated Flower model task."""
     args = _parse_args_run_flwr_model().parse_args()
 
     if not args.insecure:
@@ -64,7 +64,7 @@ def flwr_model() -> None:
 def _parse_args_run_flwr_model() -> argparse.ArgumentParser:
     """Parse `flwr-model` command line arguments."""
     parser = argparse.ArgumentParser(
-        description="Run a Flower ModelApp",
+        description="Run a Flower model task",
     )
     parser.add_argument(
         "--serverappio-api-address",
