+++ framework/py/flwr/supercore/executors/__init__.py
@@ -16,7 +16,7 @@
 
 
 from .run_agentapp import run_agentapp
-from .run_model import run_model
+from .model.run import run_model
 
 __all__ = [
     "run_agentapp",
